Refactor document~document registry inheritance
Now, the document_registry owns document rather than the other way
around.
