+++ editor/language_server/src/document_registry/document_parser.rs
